Fix formatting after invalid merge
